코드 인코딩 방식 변경

+++ Assets/Scripts/PlayerInfo.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerInfo : MonoBehaviour
 {
-    public float plAtk;
+    public int plAtk;
     public float plMoveSpd;
     public float curHp;
     public float maxHp;
